Add nickname input screen and save user's nickname.
This commit adds a new feature to the game where users can enter their nickname. The code creates a div element with an input field and submit button for the user to enter their name. If the user has already entered their name, it retrieves it from cookies. The code also saves the user's nickname in the Game class so that it can be used later on. The askforNickname() function was removed since its functionality was replaced by this new feature.